refactored AppComponent + moved code into WorkoutComponent
